Improve tests organization And bump version
